+++ src/ft_coor.c
@@ -6,7 +6,7 @@
 /*   By: yotillar <yotillar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/12 07:25:27 by yotillar          #+#    #+#             */
-/*   Updated: 2020/11/24 00:40:08 by antoine          ###   ########.fr       */
+/*   Updated: 2021/02/18 09:43:27 by antoine          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ t_coor	find_char(char **map)
 			if(map[i][j] == 'N' || map[i][j] == 'S' 
 					|| map[i][j] == 'E' || map[i][j] == 'W')
 			{
-				coor.x = (double)(j) + 0.5;
-				coor.y = (double)(i) + 0.5;
+				coor.x = (float)(j) + 0.5;
+				coor.y = (float)(i) + 0.5;
 				coor.dist = 0;
 				return (coor);
 			}
